tools/requirements.txt
openai
requests
python-dotenv
faiss-cpu
sentence-transformers
pandas
duckduckgo_search
gradio
anthropic